played around with the routes, created the route to get a single book by ID but losing the styling and JS when directed to that pathway
