+++ controllers/api/index.js
@@ -1,0 +1,6 @@
+const router = require("express").Router();
+const bookRoutes = require(`./bookRoutes`);
+
+router.use(`/book`, bookRoutes);
+
+module.exports = router;
