+++ controllers/homeRoutes.js
@@ -1,5 +1,6 @@
 const router = require("express").Router();
 const { User, Book } = require("../models");
+
 //all books
 router.get(`/`, async (req, res) => {
   try {
@@ -14,7 +15,7 @@ router.get(`/`, async (req, res) => {
   }
 });
 
-//one book by genre
+//books by genre
 router.get("/:genre", async (req, res) => {
   try {
     const booksData = await Book.findAll({
@@ -32,21 +33,22 @@ router.get("/:genre", async (req, res) => {
   } catch (err) {
     res.json(err);
   }
-
-  //one book by id
-  router.get(`/book/:id`, async (res, req) => {
-    try {
-      const bookData = await Book.findByPk(req.params.id);
-      if (!bookData) {
-        res.status(404).json({ message: "No book with this id!" });
-        return;
-      }
-      const book = bookData.get({ plain: true });
-      res.render(`book`, book);
-    } catch (err) {
-      res.status(500).json(err);
-    }
-  });
 });
+
+// //one book by id
+// router.get("/:id", async (req, res) => {
+//   try {
+//     const bookData = await Book.findByPk(req.params.id);
+//     if (!bookData) {
+//       res.status(404).json({ message: "No book with this id!" });
+//       return;
+//     }
+//     const book = bookData.get({ plain: true });
+//     console.log(book);
+//     res.render(`book`, book);
+//   } catch (err) {
+//     res.status(500).json(err);
+//   }
+// });
 
 module.exports = router;
