+++ views/home.handlebars
@@ -1,7 +1,7 @@
 <!-- featured section starts  -->
 
 <section class="featured" id="featured">
-  <h1 class="heading"><span>All Books</span></h1>
+  <h1 class="heading"><span class="path-title">All Books</span></h1>
 
   <div class="swiper featured-slider">
     <div class="swiper-wrapper">
